requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
inotify-simple==1.3.5
kornia==0.6.9
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
Pillow==9.3.0
pyqtgraph==0.13.1
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
pyzmq==24.0.1
redis==4.4.0
requests==2.28.1
shiboken6==6.4.1
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
types-Pillow==9.3.0.4
types-redis==4.3.21.7
types-setuptools==65.6.0.2
types-tabulate==0.9.0.0
types-tqdm==4.64.7.9
typing_extensions==4.4.0
urllib3==1.26.13
